% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_gene_orthologies.r
\name{build_orthology_table}
\alias{build_orthology_table}
\title{Build orthology table}
\usage{
build_orthology_table(
  taxIDs = setNames(c(9606, 9483, 10090), c("human", "marmoset", "mouse")),
  primaryTaxID = 9606,
  geneOrthologs = "https://ftp.ncbi.nlm.nih.gov/gene/DATA/gene_orthologs.gz",
  addEnsemblID = TRUE,
  addGeneInfo = c("Symbol", "description", "type_of_gene"),
  outputFilePrefix = "ortholog_table",
  returnTable = TRUE
)
}
\arguments{
\item{taxIDs}{A named vector of all species with associated taxonomy IDs (names are common names to show up in columns, and values are NCBI taxa IDs)}

\item{primaryTaxID}{A numeric vector of the species to compare against. This MUST be included in the taxIDs. Default is 9606 (human)}

\item{geneOrthologs}{The gene_orthologs.gz file updated daily. A local file in the same format can be provided, but otherwise leave the default link as is.}

\item{addEnsemblID}{Should EnsemblIDs be added for each geneID?}

\item{addGeneInfo}{Gene info to include with each gene for the primary taxonomy (NULL for none). For species other than mouse and human, this can be very slow.  Currently this is only compatible with mammalial species.  Sensible defaults are included, but all options include: "Symbol", "LocusTag", "Synonyms", "dbXrefs", "chromosome", "map_location", "description", "type_of_gene", "Symbol_from_nomenclature_authority", "Full_name_from_nomenclature_authority", "Nomenclature_status", "Other_designations", "Modification_date", "Feature_type"}

\item{outputFilePrefix}{Prefix for outputted file name (or NULL to not output file). File will be outputFilePrefix_[date].csv.}

\item{returnTable}{Should the table be returned?}
}
\value{
Table of gene orthology information including EntrezID, Ensembl ID, and gene symbol for ever available species.
}
\description{
This is the main function to call for building orthologies between any set of species
}
