% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_gene_orthologies.r
\name{homologeneOneToOne}
\alias{homologeneOneToOne}
\title{Internal function for finding one-to-one orthologs between a pair of species}
\usage{
homologeneOneToOne(tax1, tax2, orthologs)
}
\arguments{
\item{tax1}{the primary taxonomy ID}

\item{tax2}{the secondary taxonomy ID}

\item{orthologs}{An orthology table of gene IDs and taxonomy IDs}
}
\value{
a potentially reordered and filtered taxonomy table
}
\description{
Internal function for finding one-to-one orthologs between a pair of species
}
